requirements.txt
streamlit
google-genai
python-dotenv
requests
huggingface_hub
